+++ Userland/Libraries/LibWeb/CSS/Percentage.h
@@ -23,13 +23,13 @@ public:
     {
     }
 
-    explicit Percentage(float value)
+    explicit Percentage(double value)
         : m_value(value)
     {
     }
 
-    float value() const { return m_value; }
-    float as_fraction() const { return m_value * 0.01f; }
+    double value() const { return m_value; }
+    double as_fraction() const { return m_value * 0.01; }
 
     ErrorOr<String> to_string() const
     {
@@ -39,7 +39,7 @@ public:
     bool operator==(Percentage const& other) const { return m_value == other.m_value; }
 
 private:
-    float m_value;
+    double m_value;
 };
 
 }
